requirements.txt
# Multi-Agent Research & Report Writing Assistant
# Dependencies

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# LLM Providers
openai>=1.0.0
huggingface-hub>=0.20.0

# Web Framework
streamlit>=1.30.0

# Environment & Config
python-dotenv>=1.0.0

# Web Search & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Document Parsing
pdfplumber>=0.10.0
pypdf>=4.0.0

# PDF Export
fpdf2>=2.7.0

# RAG & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Metrics & Evaluation
numpy>=1.24.0

# Fine-tuning (optional - for training)
# Uncomment these to enable fine-tuning capabilities:
# peft>=0.8.0
# bitsandbytes>=0.41.0
# accelerate>=0.25.0
# datasets>=2.15.0
# trl>=0.7.0

# Optional: SerpAPI for real web search
# google-search-results>=2.4.2

